+++ app/Http/Controllers/Api/User/Admin/Guide/SpotRockController.php
@@ -8,6 +8,7 @@ use Illuminate\Http\Request;
 use App\Models\Guide\Spot_rocks_image;
 
 use App\Services\Abstract\ImageControllService;
+use App\Services\PermissionService;
 
 class SpotRockController extends Controller
 {
@@ -17,7 +18,10 @@ class SpotRockController extends Controller
     // }
 
     public function del_spot_rock_image(Request $request)
-    {   
+    {
+        $auth = PermissionService::authorize('spot_rock', 'del');
+        if ($auth) return $auth;
+        
         $image = Spot_rocks_image::where('id', '=', $request->image_id)->first();
         ImageControllService::image_delete('images/spot_rocks_img/', $image, 'image');
         $image ->delete();
