+++ app/Http/Controllers/Api/User/Admin/Shop/ShipedRegionController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 
 use App\Models\Shop\Shiped_region;
+use App\Services\PermissionService;
 
 class ShipedRegionController extends Controller
 {
@@ -15,11 +16,17 @@ class ShipedRegionController extends Controller
 
     public function get_activ_region(Request $request)
     {
+        $auth = PermissionService::authorize('shipping_region', 'view');
+        if ($auth) return $auth;
+        
         return Shiped_region::where('id',strip_tags($request->region_id))->first();
     }
 
     public function add_region(Request $request)
-    {   
+    {
+        $auth = PermissionService::authorize('shipping_region', 'add');
+        if ($auth) return $auth;
+        
         $adding_item = new Shiped_region;
 
         $adding_item['region'] = $request['adding_data']['region'];
@@ -31,6 +38,9 @@ class ShipedRegionController extends Controller
 
     public function edit_region(Request $request)
     {
+        $auth = PermissionService::authorize('shipping_region', 'edit');
+        if ($auth) return $auth;
+        
         $editing_item = Shiped_region::where('id',strip_tags($request->region_id))->first();
         
         $editing_item['region'] = $request['editing_data']['region'];
@@ -42,7 +52,11 @@ class ShipedRegionController extends Controller
 
     public function del_region(Request $request)
     {
+        $auth = PermissionService::authorize('shipping_region', 'del');
+        if ($auth) return $auth;
+        
         $deliting_item = Shiped_region::where('id',strip_tags($request->region_id))->first();
         $deliting_item ->delete();
     }
 }
+
